Refactor encryption
