+++ src/core/security/Encryption.ts
@@ -1,12 +1,7 @@
 import type { LiteralObject } from 'src/core/PDFContext';
-import type { Encrypter } from 'src/core/objects/EncryptableObject';
-import { PDFArray } from 'src/core/objects/PDFArray';
-import { PDFDict } from 'src/core/objects/PDFDict';
-import { PDFHexString } from 'src/core/objects/PDFHexString';
+import type { Encrypter } from 'src/core/objects/Encrypter';
 import type { PDFObject } from 'src/core/objects/PDFObject';
 import type { PDFRef } from 'src/core/objects/PDFRef';
-import { PDFStream } from 'src/core/objects/PDFStream';
-import { PDFString } from 'src/core/objects/PDFString';
 import type { StdSecurityHandlerDict } from 'src/core/security/StdSecurityHandler';
 import type { WordArray } from 'src/utils/crypt';
 
@@ -47,18 +42,12 @@ export abstract class EncryptionKey {
 
   /**
    * Encrypt a specified indirect object if possible.
+   * This changes the value of `indirectObject` in-place.
    */
   encryptIfPossible(indirectObject: IndirectObject): void {
     const [ref, obj] = indirectObject;
-    if (
-      obj instanceof PDFStream ||
-      obj instanceof PDFHexString ||
-      obj instanceof PDFString ||
-      obj instanceof PDFDict ||
-      obj instanceof PDFArray
-    ) {
-      indirectObject[1] = obj.encryptWith(this.createEncrypter(ref));
-    }
+    const encryptedObj = obj.encryptWith(this.createEncrypter(ref));
+    if (encryptedObj != null) indirectObject[1] = encryptedObj;
   }
 
   /**
