+++ src/core/structures/PDFContentStream.ts
@@ -1,5 +1,5 @@
 import type { PDFDict } from 'src/core/objects/PDFDict';
-import type { Encrypter } from 'src/core/objects/EncryptableObject';
+import type { Encrypter } from 'src/core/objects/Encrypter';
 import type { PDFOperator } from 'src/core/operators/PDFOperator';
 import type { PDFContext } from 'src/core/PDFContext';
 import { PDFFlateStream } from 'src/core/structures/PDFFlateStream';
